Introduce ability to add window at a specified location. The coordinate is passed in using Point2D parameter type. This allows for more complex coordinate calculations in future features, such as placement strategy.
Fixed issue with incorrect center placement.  Getting dimensions of MDIWindow was done too early. It may have been placed in Platform.runLater() but does not mean the layout bounds were actually applied yet. Therefore, mdiWindow.getLayoutBounds().getWidth() would return 0 unexpectedly.  This resulted in cases where the window was not rendered at true center of the MDI parent or was rendered at (0,0)  To ensure the width is read at correct time, run centerMdiWidth as an event listener for layoutBoundsProperty() so the calculation is done only when the data is available for sure.